feat: Created a data model form
