+++ apps/website/components/molecules/modal-right.tsx
@@ -3,8 +3,15 @@ import * as React from 'react';
 import { brandColors } from '@gateway/theme';
 
 import CloseIcon from '@mui/icons-material/Close';
-import { Stack, Typography, alpha } from '@mui/material';
-import Dialog from '@mui/material/Dialog';
+import {
+  Stack,
+  Typography,
+  alpha,
+  Paper,
+  DialogContent,
+  DialogTitle,
+} from '@mui/material';
+import Dialog, { DialogProps } from '@mui/material/Dialog';
 import IconButton from '@mui/material/IconButton';
 import Slide from '@mui/material/Slide';
 import { TransitionProps } from '@mui/material/transitions';
@@ -37,35 +44,45 @@ export default function ModalRight({
       open={open}
       onClose={handleClose}
       TransitionComponent={Transition}
+      scroll="paper"
       sx={{
         width: { xs: '100%', md: '600px', lg: '720px' },
         left: 'auto',
         right: '0',
+        '& > div > .MuiPaper-elevation': {
+          background:
+            'linear-gradient(180deg, rgba(255, 255, 255, 0.12) 0%, rgba(255, 255, 255, 0.12) 100%), #10041C',
+        },
       }}
     >
-      <Stack
+      <DialogTitle
+        component={Stack}
+        direction="row"
+        alignItems="center"
+        justifyContent="space-between"
+        sx={{
+          py: { xs: 3, md: 6 },
+          px: { xs: 3, md: 6 },
+        }}
+        id="scroll-dialog-title"
+      >
+        <Typography variant="h5">{title}</Typography>
+        <IconButton
+          aria-label="close"
+          sx={{ background: alpha(brandColors.white.main, 0.16) }}
+          onClick={() => handleClose()}
+        >
+          <CloseIcon />
+        </IconButton>
+      </DialogTitle>
+      <DialogContent
         sx={{
           px: { xs: 3, md: 6 },
-          py: { xs: 3, md: 6 },
+          pb: { xs: 3, md: 6 },
         }}
       >
-        <Stack
-          direction="row"
-          alignItems="center"
-          justifyContent="space-between"
-          sx={{ mb: 5 }}
-        >
-          <Typography variant="h5">{title}</Typography>
-          <IconButton
-            aria-label="close"
-            sx={{ background: alpha(brandColors.white.main, 0.16) }}
-            onClick={() => handleClose()}
-          >
-            <CloseIcon />
-          </IconButton>
-        </Stack>
         {children}
-      </Stack>
+      </DialogContent>
     </Dialog>
   );
 }
