+++ apps/website/components/templates/protocol/components/content-table-two-columns.tsx
@@ -1,10 +1,7 @@
-import InsertLinkIcon from '@mui/icons-material/InsertLink';
-import PhotoIcon from '@mui/icons-material/Photo';
-import PinIcon from '@mui/icons-material/Pin';
-import TextFieldsIcon from '@mui/icons-material/TextFields';
 import { Chip, Divider, Stack, Typography } from '@mui/material';
 
 import CardCell from './card-cell';
+import ChipInputType from './chip-input-type';
 
 type Props = {
   data: any[];
@@ -17,24 +14,6 @@ export default function TwoColumns({
   column2,
   isInputType = false,
 }: Props) {
-  const inputTypeContent = (inputType: string) => {
-    return (
-      <Stack
-        direction="row"
-        justifyContent="space-between"
-        alignItems="center"
-        gap={0.5}
-      >
-        {inputType === 'text' && <TextFieldsIcon sx={{ fontSize: '16px' }} />}
-        {inputType === 'number' && <PinIcon sx={{ fontSize: '16px' }} />}
-        {inputType === 'image' && <PhotoIcon sx={{ fontSize: '16px' }} />}
-        {inputType === 'link' && <InsertLinkIcon sx={{ fontSize: '16px' }} />}
-        <Typography fontSize={12} sx={{ textTransform: 'capitalize' }}>
-          {inputType}
-        </Typography>
-      </Stack>
-    );
-  };
   return (
     <Stack divider={<Divider />}>
       {data?.map((item, index) => (
@@ -49,7 +28,7 @@ export default function TwoColumns({
           </CardCell>
           <Typography sx={{ mr: { xs: 0, md: 2 }, my: 2 }}>
             {isInputType ? (
-              <Chip variant="filled" label={inputTypeContent(item[column2])} />
+              <ChipInputType type={item[column2]} />
             ) : (
               item[column2]
             )}
